[FEAT] 애플 회원탈퇴

+++ trash-app/src/main/java/com/project/trash/member/service/MemberCommandService.java
@@ -117,7 +117,11 @@ public class MemberCommandService {
   public void delete(MemberDeleteRequest param) {
     Member member = memberQueryService.getOne(MemberUtils.getMember().getSocialId());
 
-    socialApiClient.unlink(member.getSocialType(), param.getAccessToken());
+    if (member.getSocialType() == SocialType.APPLE) {
+      appleService.revoke(param.getAccessToken());
+    } else {
+      socialApiClient.unlink(member.getSocialType(), param.getAccessToken());
+    }
     member.delete();
     tokenRepository.delete(getToken(MemberUtils.getMember().getSocialId()));
   }
